Started migration of Blockly internals to typescript
